refactor: toEntity, toCondition 메서드 위치 이동
DTO -> Entity 변환 controller에서 하도록 위치 이동함
