+++ api/src/main/java/com/staybnb/rooms/service/RoomService.java
@@ -2,10 +2,7 @@ package com.staybnb.rooms.service;
 
 import com.staybnb.rooms.domain.Room;
 import com.staybnb.rooms.domain.vo.Currency;
-import com.staybnb.rooms.domain.vo.RoomType;
-import com.staybnb.rooms.dto.SearchRoomInfo;
-import com.staybnb.rooms.dto.request.CreateRoomRequest;
-import com.staybnb.rooms.dto.request.SearchRoomRequest;
+import com.staybnb.rooms.dto.SearchRoomCondition;
 import com.staybnb.rooms.dto.request.UpdateRoomRequest;
 import com.staybnb.rooms.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
@@ -24,13 +21,10 @@ public class RoomService {
 
     private final RoomRepository roomRepository;
 
-    private final UserService userService;
-    private final PlaceTypeService placeTypeService;
     private final AmenityService amenityService;
     private final ExchangeRateService exchangeRateService;
 
-    public Room save(CreateRoomRequest request) {
-        Room room = toEntity(request);
+    public Room save(Room room) {
         room.setBasePriceInUsd(exchangeRateService.convertToUSD(room.getCurrency(), room.getBasePrice()));
         return roomRepository.save(room);
     }
@@ -39,8 +33,8 @@ public class RoomService {
         return roomRepository.findById(id).orElseThrow(() -> new NoSuchElementException("존재하지 않는 숙소입니다."));
     }
 
-    public Page<Room> findAll(SearchRoomRequest request, Pageable pageable) {
-        return roomRepository.findAll(toCommand(request), pageable);
+    public Page<Room> findAll(SearchRoomCondition condition, Pageable pageable) {
+        return roomRepository.findAll(condition, pageable);
     }
 
     public Room update(long id, UpdateRoomRequest request) {
@@ -80,32 +74,4 @@ public class RoomService {
         room.setDeletedAt(LocalDateTime.now());
     }
 
-    private Room toEntity(CreateRoomRequest request) {
-        return Room.builder()
-                .host(userService.getById(request.getHostId()))
-                .placeType(placeTypeService.getByName(request.getPlaceType()))
-                .roomType(RoomType.valueOf(request.getRoomType()))
-                .address(request.getAddress())
-                .maxNumberOfGuests(request.getMaxNumberOfGuests())
-                .bedrooms(request.getBedrooms())
-                .beds(request.getBeds())
-                .amenities(amenityService.getAmenitySetByStringSet(request.getAmenities()))
-                .title(request.getTitle())
-                .description(request.getDescription())
-                .basePrice(request.getBasePrice())
-                .currency(Currency.valueOf(request.getCurrency()))
-                .build();
-    }
-
-    private SearchRoomInfo toCommand(SearchRoomRequest request) {
-        return SearchRoomInfo.builder()
-                .numberOfGuests(request.getGuests())
-                .startDate(request.getStartDate())
-                .endDate(request.getEndDate())
-                .location(request.getLocation())
-                .priceFrom(request.getPriceFrom())
-                .priceTo(request.getPriceTo())
-                .currency(request.getCurrency())
-                .build();
-    }
 }
