+++ database/src/main/java/com/staybnb/rooms/repository/custom/RoomRepositoryCustom.java
@@ -1,11 +1,11 @@
 package com.staybnb.rooms.repository.custom;
 
 import com.staybnb.rooms.domain.Room;
-import com.staybnb.rooms.dto.SearchRoomInfo;
+import com.staybnb.rooms.dto.SearchRoomCondition;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 
 public interface RoomRepositoryCustom {
-    Page<Room> findAll(SearchRoomInfo roomSearchCondition, Pageable pageable);
+    Page<Room> findAll(SearchRoomCondition cond, Pageable pageable);
 }
